requirements.txt
selenium==4.27.1
webdriver-manager==4.0.2
fastapi==0.121.0
uvicorn[standard]==0.38.0
pydantic==2.12.4
supabase==2.24.0
python-dotenv==1.1.1
psycopg2-binary==2.9.11
requests==2.32.5
